requirements.txt
pip==20.2.4
prometheus-client==0.9.0
PyYAML==5.3.1
setuptools==28.8.0
wheel==0.29.0
